beatrice/backend/pyproject.toml
[tool.poetry]
name = "beatrice"
version = "0.0.1"
description = ""
authors = ["Greta Lamù Renzi <greta@lamu.ai>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.74.1"
email-validator = "^1.1.3"
uvicorn = {extras = ["standard"], version = "^0.17.5"}
alembic = "^1.7.6"
sqlmodel = "^0.0.6"
asyncpg = "^0.25.0"
python-jose = "^3.3.0"
passlib = {extras = ["standard"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
yapf = "^0.32.0"
pylint = "^2.12.2"
pytest = "7.0.1"
pynvim = "^0.4.3"
